Fixed bug (ahem) in CheckLocality(). Added loop protection. Did pickup/dropoff.
Turns out the pickup/dropoff are extremely simple formulas. Couldn't get it working correctly though. Tracked it down to a bug in CheckLocality(). Wasn't checking for balls at all, all good now. Had to write a TestCase() for ensuring expected behavior. k1/k2 will likely need to be tuned. Also, ants refuse to walk on the edge and that's a problem.