Added ability for radius > 1
While I got the radius to work for sizes bigger than one it fails to clump larger. This is likely due to the probability for pickup/dropoff. I made loads of changes but got tired of tweaking it. It works as pushed and will always work on radius 1. Radius 2-5 seem to work but do not make larger clusters.